+++ lizard_progress/exports.py
@@ -101,6 +101,7 @@ def export_as_metfile(export_run):
 
     want_sorted_measurements = errors.ErrorConfiguration(
         project=export_run.project,
+        organization=None,  # Give either project or organization
         measurement_type=models.AvailableMeasurementType.dwarsprofiel()
         ).wants_sorted_measurements()