+++ lizard_progress/parsers/met_parser.py
@@ -34,7 +34,7 @@ class MetParser(specifics.ProgressParser):
 
     def parse(self, check_only=False):
         self.error_config = (
-            errors.ErrorConfiguration(self.project, self.mtype()))
+            errors.ErrorConfiguration(self.project, None, self.mtype()))
 
         parsed_metfile = parse_metfile(self.file_object)